Physics2D.Net: added: CreateFromBitmap to the Polygon class a BitmapHelper class that assists it.
AdvanceMath:
changed: the contains methods to be closer to XNAs model.
fixed: an elusive bug with the BoundingPolygon.GetDistance method and it should be faster (in theory).
changed: LineSegment.Getdistance so it will no longer return a negative value.
changed: the order of parameters to the BoundingRectagle's constructors.
removed: origin from Vector2D use Zero instead.