AdvanceMath changed: Vector2D.GetAngle to use a more accepted approuch.
Physics2D.Net
BREAKING: Changed: renamed all the ignorers to not hav eth word Collision in front of them.
Added: GroupCollection. which holds most of the logic GroupIgnorer did.
BREAKING: Changed: GroupIgnorer so that it uses the new GroupsCollection.
BREAKING: Changed: renamed Body.Ignorer to Body.CollisionIgnorer
Added: AdvGroupIgnorer which has 2 GroupCollections one of the group its part of one of the groups it ignores.
Added: Body.EventIgnorer. this is to allow ignoring of collision events